safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A8_T1-expected.txt

CONSOLE MESSAGE: line 86: SyntaxError: Parse error
S12.7_A8_T1

PASS Expected parsing failure

TEST COMPLETE
